+++ packages/expo-network/mocks/ExpoNetwork.ts
@@ -1,0 +1,9 @@
+export async function getNetworkStateAsync(): Promise<any> {}
+
+export async function getIpAddressAsync(): Promise<any> {}
+
+export async function isAirplaneModeEnabledAsync(): Promise<any> {}
+
+export async function getCellSignalStrengthAsync(): Promise<any> {}
+
+export async function getWifiSignalStrengthAsync(): Promise<any> {}